+++ packages/cli/test/cli.test.ts
@@ -1,6 +1,7 @@
 import { describe, it, beforeEach, afterEach } from 'mocha';
 import * as expect from 'expect';
 import * as process from 'process';
+import { promises as fs } from 'fs';
 
 import { Process } from './helpers/process';
 import { World } from './helpers/world';
@@ -227,6 +228,99 @@ describe('@bigtest/cli', function() {
         expect(child.stdout?.output).toContain('Cannot run tests due to build errors in the test suite')
         expect(child.stdout?.output).toContain('test/fixtures/bad.broken.ts')
         expect(child.stdout?.output).toContain('⨯ FAILURE')
+      });
+    });
+  });
+
+  describe('coverage', () => {
+    beforeEach( async() => {
+      await fs.rmdir('./tmp/coverage', { recursive: true });
+    });
+
+    describe('when coverage output is requested', () => {
+      let child: Process;
+      beforeEach(async () => {
+        child = await World.spawn(run('ci', '--coverage', './test/fixtures/coverage.test.ts'));
+        await World.spawn(child.join());
+      });
+
+      it('outputs the coverage reports', async () => {
+        await expect((fs.access('tmp/coverage/lcov/lcov.info'))).resolves.toBeUndefined();
+        expect(child.stdout?.output).toContain('reported to -> ./tmp/coverage');
+      });
+    });
+
+    describe('when requested, but there is no coverage data in the test run', () => {
+      let child: Process;
+      beforeEach(async () => {
+        child = await World.spawn(run('ci', '--coverage', './test/fixtures/passing.test.ts'));
+        await World.spawn(child.join());
+      });
+
+      it('warns the user that there is no coverage data', () => {
+        expect(child.stderr?.output).toContain('no coverage metrics were present');
+      });
+    });
+
+    describe('when coverage data is present, but coverage output is not requested', () => {
+      let child: Process;
+      beforeEach(async () => {
+        child = await World.spawn(run('ci', './test/fixtures/coverage.test.ts'));
+        await World.spawn(child.join());
+      });
+
+      it('does nothing', async () => {
+        expect(child.stdout?.output).not.toContain('@bigtest/coverage');
+        await expect((fs.access('tmp/coverage/lcov/lcov.info'))).rejects.toMatchObject({
+          message: expect.stringContaining('no such file or directory')
+        });
+      });
+    });
+  });
+
+  describe('init', () => {
+    describe('running the init command', () => {
+      let child: Process;
+
+      beforeEach(async () => {
+        child = await World.spawn(run('init', '--config-file', './tmp/bigtest-config-test.json'));
+
+        await World.spawn(child.stdout?.waitFor('Which port would you like to run BigTest on?'));
+        child.stdin?.write('not-a-port\n');
+        await World.spawn(child.stdout?.waitFor('Not a number!'));
+        child.stdin?.write('1234\n');
+
+        await World.spawn(child.stdout?.waitFor('Where are your test files located?'));
+        child.stdin?.write('test.ts\n');
+
+        await World.spawn(child.stdout?.waitFor('Do you want BigTest to start your application for you?'));
+        child.stdin?.write('\n');
+
+        await World.spawn(child.stdout?.waitFor('What command do you run to start your application?'));
+        child.stdin?.write('yarn run-my-app\n');
+
+        await World.spawn(child.stdout?.waitFor('Which port would you like to run your application on?'));
+        child.stdin?.write('9000\n');
+
+        await World.spawn(child.stdout?.waitFor('Which URL do you use to access your application?'));
+        child.stdin?.write('\n');
+
+        await World.spawn(child.join());
+      });
+
+      afterEach(async () => {
+        await World.spawn(child.close());
+      });
+
+      it('exits successfully and writes a new config file', async () => {
+        expect(child.code).toEqual(0);
+        let buffer = await fs.readFile('./tmp/bigtest-config-test.json');
+        let config = JSON.parse(buffer.toString());
+        expect(config.port).toEqual(1234);
+        expect(config.testFiles).toEqual(['test.ts']);
+        expect(config.app.command).toEqual('yarn run-my-app');
+        expect(config.app.env.PORT).toEqual(9000);
+        expect(config.app.url).toEqual('http://localhost:9000');
       });
     });
   });