+++ packages/cli/test/helpers/process.ts
@@ -12,6 +12,7 @@ interface ProcessOptions {
 export class Process {
   public stdout?: Stream;
   public stderr?: Stream;
+  public stdin?: ChildProcess['stdin'];
   private child?: ChildProcess;
   private exited = Deferred<[number,number]>();
 
@@ -84,6 +85,9 @@ export class Process {
     }
     if (this.child.stderr) {
       this.stderr = yield Stream.of(this.child.stderr, this.options.verbose);
+    }
+    if (this.child.stdin) {
+      this.stdin = this.child.stdin;
     }
 
     let [code, signal] = yield once(this.child, 'exit');