+++ packages/interactor/src/definitions/heading.ts
@@ -1,4 +1,5 @@
 import { createInteractor } from '../create-interactor';
+import { isVisible } from 'element-is-visible';
 
 export const Heading = createInteractor<HTMLHeadingElement>('heading')({
   selector: 'h1,h2,h3,h4,h5,h6',
@@ -6,6 +7,7 @@ export const Heading = createInteractor<HTMLHeadingElement>('heading')({
     byId: (element) => element.id
   },
   filters: {
-    level: (element) => parseInt(element.tagName[1])
+    level: (element) => parseInt(element.tagName[1]),
+    visible: { apply: isVisible, default: true },
   }
 });