+++ packages/interactor/test/definitions/heading.test.ts
@@ -27,5 +27,29 @@ describe('@bigtest/interactor', () => {
         await expect(Heading('Foo', { level: 3 }).exists()).rejects.toHaveProperty('name', 'NoSuchElementError');
       });
     });
+
+    describe('filter `visible`', () => {
+      it('filters heading tags by their visibility', async () => {
+        dom(`
+          <h1 style="display:none;">Foo</h1>
+          <h2>Bar</h2>
+        `);
+
+        await expect(Heading('Foo', { visible: false }).exists()).resolves.toBeUndefined();
+        await expect(Heading('Bar', { visible: true }).exists()).resolves.toBeUndefined();
+        await expect(Heading('Foo', { visible: true }).exists()).rejects.toHaveProperty('name', 'NoSuchElementError');
+        await expect(Heading('Bar', { visible: false }).exists()).rejects.toHaveProperty('name', 'NoSuchElementError');
+      });
+
+      it('defaults to `true`', async () => {
+        dom(`
+          <h1 style="display:none;">Foo</h1>
+          <h2>Bar</h2>
+        `);
+
+        await expect(Heading('Bar').exists()).resolves.toBeUndefined();
+        await expect(Heading('Foo').exists()).rejects.toHaveProperty('name', 'NoSuchElementError');
+      });
+    });
   });
 });