+++ packages/server/src/result-aggregator/test-run.ts
@@ -4,6 +4,7 @@ import { TestRunState } from '../orchestrator/state';
 import { Aggregator, AggregatorOptions } from './aggregator';
 import { TestRunAgentAggregator } from './test-run-agent';
 import { parallel } from './parallel';
+import { createCoverageMap, CoverageMap, CoverageMapData } from 'istanbul-lib-coverage';
 
 export class TestRunAggregator extends Aggregator<TestRunState, AggregatorOptions> {
   *perform(): Operation<ResultStatus> {
@@ -14,9 +15,21 @@ export class TestRunAggregator extends Aggregator<TestRunState, AggregatorOption
     });
 
     let statuses: ResultStatus[] = yield parallel(agentRuns);
+    let status: ResultStatus = statuses.some(status => status === 'failed') ? 'failed' : 'ok';
 
-    this.statusSlice.set(statuses.some(status => status === 'failed') ? 'failed' : 'ok');
+    let coverage = Object.values(this.slice.get().agents)
+      .filter(agent => !!agent.coverage)
+      .map(agent => agent.coverage as CoverageMapData)
+      .reduce((current: CoverageMap | undefined, data) => {
+        if (!current) {
+          return createCoverageMap(data);
+        } else {
+          current.merge(data);
+          return current;
+        }
+      }, undefined)
 
-    return this.statusSlice.get();
+    this.slice.update(state => ({ ...state, status, coverage }));
+    return status;
   }
 }