+++ packages/server/test/helpers.ts
@@ -79,7 +79,8 @@ export const actions = {
             port: 24103,
           },
           drivers: {},
-          launch: []
+          launch: [],
+          coverage: { reports: [], directory: "" }
         }
       }));