+++ packages/suite/test/dsl.test.ts
@@ -1,7 +1,10 @@
 import { describe, it } from 'mocha';
 import * as expect from 'expect'
 
+import { test }from '../src';
 import example from './fixtures/example';
+
+const noop = () => undefined;
 
 describe('dsl', () => {
   it('returns a serialized test suite', async () => {
@@ -16,5 +19,22 @@ describe('dsl', () => {
     expect(example.children[0].assertions[0].description).toEqual('a child assertion');
 
     await expect(example.steps[0].action({})).resolves.toHaveProperty('foo', 'foo');
+  });
+
+  it('can have multiple steps', () => {
+    let { steps } = test('a test')
+      .step(
+        { description: "hello", action: noop },
+        { description: "world", action: noop });
+
+    expect(steps.map(step => step.description)).toEqual(['hello', 'world']);
+  });
+  it('can have multiple assertions', () => {
+    let { assertions } = test('an assertion').
+      assertion(
+        { description: "hello", check: noop },
+        { description: "world", check: noop });
+
+    expect(assertions.map(assertion => assertion.description)).toEqual(['hello', 'world']);
   });
 })