+++ packages/project/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @bigtest/project
 
+## 0.11.0
+
+### Minor Changes
+
+- 4b54d9f9: Add an interactive `init` command
+
+## 0.10.0
+
+### Minor Changes
+
+- 248d6ddc: produce coverage reports by passing the `--coverage` option to the
+  `test` and `ci` commands
+
 ## 0.9.0
 
 ### Minor Changes