+++ packages/server/CHANGELOG.md
@@ -1,5 +1,31 @@
 # Changelog
 
+## 0.16.1
+
+### Patch Changes
+
+- Updated dependencies [4b54d9f9]
+  - @bigtest/project@0.11.0
+  - @bigtest/bundler@0.10.2
+
+## 0.16.0
+
+### Minor Changes
+
+- 248d6ddc: produce coverage reports by passing the `--coverage` option to the
+  `test` and `ci` commands
+
+### Patch Changes
+
+- Updated dependencies [248d6ddc]
+- Updated dependencies [4486733a]
+- Updated dependencies [f51c9933]
+  - @bigtest/project@0.10.0
+  - @bigtest/agent@0.14.0
+  - @bigtest/atom@0.10.0
+  - @bigtest/bundler@0.10.1
+  - @bigtest/webdriver@0.6.3
+
 ## 0.15.0
 
 ### Minor Changes