+++ packages/webdriver/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @bigtest/webdriver
 
+## 0.6.3
+
+### Patch Changes
+
+- Updated dependencies [f51c9933]
+  - @bigtest/atom@0.10.0
+
 ## 0.6.2
 
 ### Patch Changes